Create Output results file
This is the output results file that represent one of the simulations in NetLogo. The code can be runned as a whole, changing the filepath of the simulation you look at. The file names must also change to fit your directory. 
This was used for all 12 simulations ( 3 replicates for each area) where the csv files for season was further used to make results plot.