Create Interaction, reluse,relcatch plot
This scripts describe how the plotting of results were conducted. The map of Norway with interaction, use and catch plot are based on a combined dataset made in Excel based on the mean value of all seasons files. The script also include a zoomed in plot of the study area of Sør-Trøndelag. Much of the code can be used as they are, only changing file paths and have the previous (location, output files) read in